Refactor our channel to release version parsing logic
Splits up `ReleaseManifest` into 3 pieces:
- A. Downloading of .NET Archives
- B. Accessing the release manifest & release manifest data structures
- C. Parsing a 'channel' into a ReleaseVersion

This is a better separation of responsibility and makes each class a far more reasonable size.
