+++ src/Installer/Microsoft.Dotnet.Installation/Internal/DotnetReleaseInfoProvider.cs
@@ -13,7 +13,7 @@ internal class DotnetReleaseInfoProvider : IDotnetReleaseInfoProvider
     public IEnumerable<string> GetAvailableChannels() => throw new NotImplementedException();
     public ReleaseVersion? GetLatestVersion(InstallComponent component, string channel)
     {
-        var releaseManifest = new ReleaseManifest();
+        var releaseManifest = new ChannelVersionResolver();
         var release = releaseManifest.GetLatestVersionForChannel(new UpdateChannel(channel), component);
 
         return release;
